+++ src/io/sys/windows/pipe/pipe_write.rs
@@ -25,7 +25,7 @@ impl<'a> PipeWrite<'a> {
     ) -> Self {
         let pipe = s.as_raw_handle();
         PipeWrite {
-            io_data: EventData::new(pipe as isize),
+            io_data: EventData::new(pipe),
             buf,
             pipe,
             #[cfg(feature = "io_timeout")]
